requirements.txt
langchain>=0.1.15
langchain-core>=0.1.33
langchain-community>=0.0.19
langchainhub
openai>=1.13.3
transformers>=4.40.0
accelerate
huggingface_hub
sentence-transformers
mistral
scikit-learn
pymupdf
pdfminer.six
python-docx
unstructured
docx2txt
faiss-cpu
chromadb
pymongo
motor
tiktoken
fastapi
uvicorn[standard]
python-multipart
aiofiles
python-dotenv
fastapi-cors
pandas
nltk
httpx
regex
pytest
